+++ changelog.md
@@ -7,8 +7,8 @@
 
 - Add `click:source` directive as an alias to `click:example` directive in Sphinx extension.
 - Flag `click:example` directive as deprecated in favor of `click:source`.
+- Add support for nested GitHub alerts.
 - Fix fetching version when the CLI is implemented as a standalone script and not as a package module.
-- Fix conversion of GitHub alerts in nested admonitions.
 
 ## [7.2.0 (2025-11-26)](https://github.com/kdeldycke/click-extra/compare/v7.1.0...v7.2.0)